added config structs and migrated lib.rs code to separate modules
